mres register is 4bits

+++ src/TMC2208Stepper.h
@@ -124,7 +124,7 @@ class TMC2208Stepper {
 		void tbl(uint8_t B);
 		void blank_time(uint8_t B);
 		void vsense(bool B);
-		void mres(uint16_t B);
+		void mres(uint8_t B);
 		void intpol(bool B);
 		void dedge(bool B);
 		void diss2g(bool B);
